[FORMATTER] [NINJS]: added time to read to output
time to read (estimated number of minute to read body of item) is added
in readTime property. The property has been added to Superdesk ninjs
schema too.

moved some methods from etree module to the new text_utils

SDFID-118
