+++ apps/templates/__init__.py
@@ -10,7 +10,7 @@
 
 import superdesk
 from superdesk import register_jinja_filter
-from superdesk.etree import get_text
+from superdesk.text_utils import get_text
 from .content_templates import ContentTemplatesResource, ContentTemplatesService, CONTENT_TEMPLATE_PRIVILEGE
 from .content_templates import ContentTemplatesApplyResource, ContentTemplatesApplyService
 from .content_templates import create_scheduled_content  # noqa
